implement smart updating

+++ build/universe.js
@@ -1,6 +1,21 @@
 import fs from "fs-extra";
 import { stringify } from "./json.js";
 import sharp from "sharp";
+import { createHash } from "crypto"
+
+async function computeNodeSHA1(directory) {
+  const sha1 = createHash('sha1');
+  
+  let blob = "";
+  blob += await fs.readFile(directory + "/node.json");
+  blob += await fs.readFile(directory + "/node.png");
+  
+  for(const file of await fs.readdir(directory + "/lang")) {
+    blob += await fs.readFile(directory + "/lang/" + file);
+  }
+
+  return sha1.update(blob).digest('hex');
+}
 
 export async function buildUniverse() {
   await fs.remove("dist");
@@ -14,9 +29,45 @@ export async function buildUniverse() {
   await fs.mkdirs("dist/universe/lang");
   await fs.mkdirs("dist/universe/assets");
 
+  await fs.mkdirs(".cache");
+  await fs.mkdirs(".cache/checksums");
+  await fs.mkdirs(".cache/universe-nodes");
+  await fs.mkdirs(".cache/lang-nodes");
+
   const universeNodeIds = await fs.readdir("universe");
   await Promise.all(
     universeNodeIds.map(async (id) => {
+      if(id == ".DS_Store") return;
+      const checksumFileExists = await fs.pathExists(`.cache/checksums/${id}.sha1`);
+      if(checksumFileExists) {
+        const checksum = await fs.readFile(`.cache/checksums/${id}.sha1`);
+        if(checksum.toString() === await computeNodeSHA1(`universe/${id}`)) {
+          const universeNode = JSON.parse(await fs.readFile(`.cache/universe-nodes/${id}.json`));
+          universeNodes[id] = universeNode;
+
+          const langNodeIds = (await fs.readdir(`universe/${id}/lang`)).map(langId => langId.replace(/\.md$/, ""));
+          await Promise.all(
+            langNodeIds.map(async (langId) => {
+              const langNode = JSON.parse(await fs.readFile(`.cache/lang-nodes/${id}-${langId}.json`));
+              languages[langId] = languages[langId] || {};
+              languages[langId][id] = langNode;
+            })
+          );
+
+          await fs.copy(
+            `universe/${id}/node.png`,
+            `dist/universe/assets/${id}.png`
+          );
+
+          return;
+        } else {
+          console.log(`Change detected: ${id}`);
+          await fs.remove(`.cache/checksums/${id}.sha1`);
+        }
+      } else {
+        console.log(`New node detected: ${id}`);
+      }
+
       const node = await fs.readJSON(`universe/${id}/node.json`);
       universeNodes[id] = node;
 
@@ -29,6 +80,8 @@ export async function buildUniverse() {
 
         const image = sharp(`universe/${id}/node.png`);
         const meta = await image.metadata();
+        universeNodes[id].width = meta.width;
+        universeNodes[id].height = meta.height;
         image.resize(Math.ceil(meta.width / 4), Math.ceil(meta.height / 4));
         image.stats()
           .then(({ channels: [rc, gc, bc] }) => {
@@ -42,6 +95,8 @@ export async function buildUniverse() {
       } else {
         console.log(`No node image for ${id}`);
       }
+
+      fs.writeFile(`.cache/universe-nodes/${id}.json`, stringify(universeNodes[id]));
 
       const langs = await fs.readdir(`universe/${id}/lang`);
       await Promise.all(
@@ -62,8 +117,12 @@ export async function buildUniverse() {
 
           languages[lang] = languages[lang] || {};
           languages[lang][id] = langNode;
+
+          fs.writeFile(`.cache/lang-nodes/${id}-${lang}.json`, stringify(langNode));
         })
       );
+
+      fs.writeFile(`.cache/checksums/${id}.sha1`, await computeNodeSHA1(`universe/${id}`));
     })
   );
 
@@ -71,10 +130,17 @@ export async function buildUniverse() {
 
   await Promise.all(
     Object.keys(languages).map(async (id) => {
-      await fs.writeFile(
-        `dist/universe/lang/${id}.json`,
-        stringify(languages[id])
-      );
+      if(!(await fs.pathExists(`dist/universe/lang/${id}.json`))) {
+        await fs.writeFile(
+          `dist/universe/lang/${id}.json`,
+          stringify(languages[id])
+        );
+      } else {
+        const oldLang = JSON.parse(await fs.readFile(`dist/universe/lang/${id}.json`))
+        for(const nodeId of Object.keys(languages[id])) {
+          oldLang[nodeId] = languages[id][nodeId];
+        }
+      }
     })
   );
